Apparish slight modifications
I have changed some bits of apparish so that they make the shellcheck
utility happy, which is probably at least a bit of a facsimile for good
shell code.

I have also tried to make the style of indentation and function
definition a little more consistent, for mostly selfish reasons.

The completion registry is now done through arrays, which are defined in
only one place, which is a little easier to maintain, I think.